moved apiUtils to server directory
