+++ src/server/utils/apiUtils/timestamps.test.js
@@ -1,4 +1,3 @@
-import dateFormat from 'dateformat';
 import { expect } from 'chai';
 
 import {
